Version 97.2 (delete page with missing data store)
modified:   handlers/handle,admin,pages,api,permanently-delete-from-the-trash.php
-   When a data store doesn't exist for a page it shouldn't prevent the delete from succeeding.

+++ handlers/handle,admin,pages,api,permanently-delete-from-the-trash.php
@@ -44,7 +44,17 @@ Colby::mysqli()->autocommit(false);
 CBPages::deleteRowWithDataStoreIDFromTheTrash($dataStoreID);
 
 $dataStore = new CBDataStore($dataStoreID);
-$dataStore->delete();
+
+/**
+ * In most cases the data store directory can be assumed to exists but there
+ * are rare scenarios where it won't and deleting a page shouldn't fail in those
+ * scenarios.
+ */
+
+if (is_dir($dataStore->directory()))
+{
+    $dataStore->delete();
+}
 
 Colby::mysqli()->commit();
 Colby::mysqli()->autocommit(true);
